fix: rename scss module
